+++ src/main/java/edu/metrostate/booknow/Utils/AlertUtil.java
@@ -1,0 +1,29 @@
+package edu.metrostate.booknow.Utils;
+
+import javafx.scene.control.Alert;
+
+/**
+ * Utility class for displaying different types of alert dialogs.
+ */
+
+public class AlertUtil {
+
+    // Displays information alert dialog
+    public static void showInfoAlert(String title, String message) {
+        Alert alert = new Alert(Alert.AlertType.INFORMATION);
+        alert.setTitle(title);
+        alert.setHeaderText(null);
+        alert.setContentText(message);
+        alert.showAndWait();
+    }
+
+    // Displays error alert dialog
+    public static void showErrorAlert(String title, String message) {
+        Alert alert = new Alert(Alert.AlertType.ERROR);
+        alert.setTitle(title);
+        alert.setHeaderText(null);
+        alert.setContentText(message);
+        alert.showAndWait();
+    }
+
+}