+++ src/main/java/module-info.java
@@ -2,6 +2,7 @@ module edu.metrostate.booknow {
     requires javafx.controls;
     requires javafx.fxml;
     requires java.sql;
+    requires java.desktop;
 
 
     opens edu.metrostate.booknow to javafx.fxml;